+++ src/helpers.js
@@ -97,8 +97,8 @@ export function constructInstructions(headline, description) {
   return `
 <div class="calibration-instruction">
   <h1>${headline}</h1>
-  <p>
-${description}
+  <p class="calibration-description">
+    ${description}
   </p>
 </div>`
 }